fix(delete): handle stale pubky detail route after delete
